Fixed some errors in the BingMaps providers classes: changed Integers to Doubles

+++ applicatie/VerkeerDomain/src/main/java/be/ugent/verkeer4/verkeerdomain/provider/BingMapsProvider.java
@@ -48,7 +48,7 @@ public class BingMapsProvider extends BaseProvider implements IProvider {
                 
                 RouteData rd = setRouteData(route,
                         distance,
-                        (int) bingResource.getTravelDurationTraffic(),
+                        (int) ((double) bingResource.getTravelDurationTraffic()),
                         delay);
 
                 return rd;
